feat: support attach in sqlite
