+++ test/sqlite.spec.js
@@ -9,8 +9,13 @@ describe('sqlite', () => {
 		return parser.sqlify(ast, opt);
   }
 
-	it('should support analyze in sqlite', () => {
+	it('should support analyze', () => {
 		const sql = 'analyze schemaName.tableName'
 		expect(getParsedSql(sql)).to.be.equal('ANALYZE `schemaName`.`tableName`')
+	})
+
+	it('should support attach', () => {
+		const sql = "attach database 'c:\sqlite\db\contacts.db' as contacts;"
+		expect(getParsedSql(sql)).to.be.equal("ATTACH DATABASE 'c:\sqlite\db\contacts.db' AS `contacts`")
 	})
 })
